[Enhancement] NavigationTest Helper
