+++ app/src/test/java/com/hello/curiosity/navigation/CuriosityNavHostTest.kt
@@ -1,35 +1,25 @@
 package com.hello.curiosity.navigation
 
-import androidx.compose.ui.platform.LocalContext
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.navigation.compose.ComposeNavigator
-import androidx.navigation.testing.TestNavHostController
-import com.hello.curiosity.test.compose.ComposeTest
+import com.hello.curiosity.test.compose.NavigationTest
 import com.hello.curiosity.ui.scenes.Scenes
 import org.junit.Assert.assertEquals
-import org.junit.Before
 import org.junit.Test
 
-class CuriosityNavHostTest : ComposeTest() {
+class CuriosityNavHostTest : NavigationTest() {
 
-    private lateinit var navController: TestNavHostController
-
-    @Before
-    fun setup() {
-        composeTestRule.setContent {
-            navController = TestNavHostController(LocalContext.current)
-            navController.navigatorProvider.addNavigator(ComposeNavigator())
-            CuriosityNavHost(navController = navController)
-        }
-
-        composeTestRule
-            .onNodeWithTag("component-scene-test-tag")
-            .assertIsDisplayed()
+    override val navHost: @Composable () -> Unit = {
+        CuriosityNavHost(navController = navController)
     }
 
     @Test
     fun `verify start destination is ComponentScene`() {
+        composeTestRule
+            .onNodeWithTag("component-scene-test-tag")
+            .assertIsDisplayed()
+
         val route = navController.currentBackStackEntry?.destination?.route
         assertEquals(route, Scenes.Components.route)
     }
@@ -88,11 +78,5 @@ class CuriosityNavHostTest : ComposeTest() {
         composeTestRule
             .onNodeWithTag("type-scene-test-tag")
             .assertIsDisplayed()
-    }
-
-    private fun assertNavigateToDestination(scene: String) {
-        navController.navigate(scene)
-        val route = navController.currentBackStackEntry?.destination?.route
-        assertEquals(route, scene)
     }
 }
